# contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ТехноСпортФест 2026 API",
    "version": "1.0.0",
    "description": "Канонический контракт MVP. Времена UTC. JSON camelCase. Неизвестные поля запросов отклоняются. Списки сортируются детерминированно: competitions по startsAt,id; participants по fullName,registrationId; results по place,fullName,id; news/documents по publishedAt DESC,id DESC; rating по points DESC,fullName,id."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Зарегистрировать спортсмена и войти",
        "operationId": "register",
        "x-owner": "B1",
        "x-milestone": "v1",
        "responses": {
          "201": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "tsf_access=<JWT>; HttpOnly; SameSite=Lax; Secure в production; Path=/api/v1; Max-Age=86400",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "description": "Роль всегда athlete. Email уникален без учёта регистра. Выдаёт JWT на 24 часа в cookie.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "x-origin-check": true
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Войти",
        "operationId": "login",
        "x-owner": "B1",
        "x-milestone": "v1",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "tsf_access=<JWT>; HttpOnly; SameSite=Lax; Secure в production; Path=/api/v1; Max-Age=86400",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "description": "Неверная пара email/пароль возвращает 401 с code INVALID_CREDENTIALS. Выдаёт JWT на 24 часа в cookie.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "x-origin-check": true
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Выйти",
        "operationId": "logout",
        "x-owner": "B1",
        "x-milestone": "v1",
        "responses": {
          "204": {
            "description": "Cookie очищена. Уже выданный JWT нельзя отозвать до exp.",
            "headers": {
              "Set-Cookie": {
                "description": "tsf_access=; HttpOnly; SameSite=Lax; Secure в production; Path=/api/v1; Max-Age=0",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        },
        "security": [],
        "x-origin-check": true
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Текущий пользователь",
        "operationId": "getCurrentUser",
        "x-owner": "B1",
        "x-milestone": "v1",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        },
        "security": [
          {
            "accessCookie": []
          }
        ]
      }
    },
    "/me": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Мой профиль",
        "operationId": "getMyProfile",
        "x-owner": "B1",
        "x-milestone": "v2",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        },
        "security": [
          {
            "accessCookie": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Profile"
        ],
        "summary": "Обновить мой профиль",
        "operationId": "updateMyProfile",
        "x-owner": "B1",
        "x-milestone": "v2",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "description": "Только athlete. disciplineIds — полный новый набор без дублей; все ID должны существовать.",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatch"
              }
            }
          }
        },
        "x-origin-check": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "accessCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tsf_access",
        "description": "JWT, HttpOnly; SameSite=Lax; Secure в production; срок 24 часа"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code"
          },
          "message": {
            "type": "string",
            "description": "Human-readable Russian message"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Optional field errors"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "User": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "athlete",
              "organizer"
            ],
            "description": "Имя роли пользователя: системные athlete/organizer/master-admin или динамическая роль master admin"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Права текущей роли, например news.create"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "createdAt",
          "permissions"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128
          },
          "fullName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 150
          }
        },
        "required": [
          "email",
          "password",
          "fullName"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "Profile": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "fullName": {
            "type": "string"
          },
          "education": {
            "type": "string",
            "maxLength": 200,
            "nullable": true
          },
          "locality": {
            "type": "string",
            "maxLength": 150,
            "nullable": true
          },
          "disciplineIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "user",
          "fullName",
          "education",
          "locality",
          "disciplineIds"
        ]
      },
      "ProfilePatch": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 150
          },
          "education": {
            "type": "string",
            "maxLength": 200,
            "nullable": true
          },
          "locality": {
            "type": "string",
            "maxLength": 150,
            "nullable": true
          },
          "disciplineIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "minProperties": 1
      },
...
    },
    "responses": {
...
      "401": {
        "description": "Нет действующего JWT",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "UNAUTHENTICATED",
              "message": "Нет действующего JWT"
            }
          }
        }
      },
      "403": {
        "description": "Недостаточно прав или неверный/отсутствующий Origin",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "FORBIDDEN",
              "message": "Недостаточно прав или неверный/отсутствующий Origin"
            }
          }
        }
      },
...
      "409": {
        "description": "Конфликт состояния или уникальности",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "CONFLICT",
              "message": "Конфликт состояния или уникальности"
            }
          }
        }
      },
      "422": {
        "description": "Ошибка полей",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "code": "VALIDATION_ERROR",
              "message": "Ошибка полей"
            }
          }
        }
      }
    }
  }
}